requirements.txt
boto3==1.34.15
azure-storage-blob==12.9.0
python-dotenv==1.0.1